requirements.txt
streamlit
kiteconnect
pandas
python-dotenv
requests
schedule